+++ docs/source/version_history.rst
@@ -2,8 +2,11 @@
 Version History
 ===============
 
-V2.5.0 - 2021 / XX / XX
+V2.5.0 - 2022 / 02 / XX
 =======================
+
+* Field classes no longer have :code:`flip=True` as the default - it is now :code:`False`.
+  Please check any field maps created by scripts using these classes.
 
 New Features
 ------------
@@ -21,6 +24,9 @@ New Features
   an array is written in. This can be used to validate any field manipulations.
 * New field plotting function `pybdsim.Field.Plot1DFxFyFz` to see field components in 1D.
 * Field loading automatically works for dimensions such as X, Z for 2D instead of X, Y now.
+* Ability to load a rebdsim output file and only load the ROOT histograms without loading
+  the BDSIM and rebdsim shared libraries, so it can be used on a separate computer with just
+  ROOT.
 
 Bug Fixes
 ---------
@@ -30,10 +36,12 @@ Bug Fixes
   This has been fixed by fixing the interpretation of the error codes from ROOT.
 * Fix warning about "nonposy" in matplotlib version for log scales.
 * Fix check in Run of if it's a ROOT file or not. Simplify it to use file extension.
+* Tolerate no pytransport installation.
 
 General
 -------
 
+* Updated out of date documentation.
 * Better automatic ranges for Histogram1DMultiple plots by default.
 * Better field loading in `pybdsim.Field.Load`. Returns the same Field object
   from pybdsim as you would write.